+++ CHANGELOG.md
@@ -12,8 +12,11 @@ I understand how Elixir uses it.
 
 ### Added
 
-- Lens2.Lenses.Keyword lens makers
+- `Lens2.Lenses.Keyword` lens makers
 - Lenses for `BiMap` now also handle `BiMultiMap`.
+- `defmakerp` and `deflensp`
+– Because [TypedStructLens](https://hexdocs.pm/typed_struct_lens/readme.html) has a
+  hardcoded dependency on Lens 1, I had to make a copy to alias. 
 
 ### Changed
 - `Lens.Lenses.BiMap` changed to `Lens.Lenses.Bi` because it now handles both